[#18] Refactor: 프로젝트 구조를 DDD(Domain-Driven Design) 패턴으로 변경
- 도메인별 패키지 분리 (Player, Team, CheerSong, Game, Lineup)
- 각 도메인은 domain, application, infrastructure 레이어로 구성
- 공통 컴포넌트는 shared 패키지로 분리
- Clean Architecture 원칙에 따른 의존성 방향 설정
- 도메인 간 느슨한 결합으로 확장성 및 유지보수성 향상
- CLAUDE.md 파일 추가로 DDD 구조 문서화
- logs/ 디렉토리를 .gitignore에 추가
