--- src/main/java/academy/cheerlot/domain/CheerSong.java
@@ -1,27 +1,0 @@
-package academy.cheerlot.domain;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "cheer_song")
-@Data
-@NoArgsConstructor
-public class CheerSong {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    private String title;
-    
-    @Column(columnDefinition = "TEXT")
-    private String lyrics;
-    
-    private String audioFileName;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "player_id")
-    private Player player;
-}
