--- src/main/java/academy/cheerlot/domain/Game.java
@@ -1,15 +1,0 @@
-package academy.cheerlot.domain;
-
-import jakarta.persistence.*;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Entity
-@Table(name = "game")
-@Data
-@NoArgsConstructor
-public class Game {
-
-    @Id
-    private String gameId;
-}
