--- src/main/java/academy/cheerlot/dto/LineupResponse.java
@@ -1,10 +1,0 @@
-package academy.cheerlot.dto;
-
-import java.util.List;
-
-public record LineupResponse(
-        String updated,
-        String opponent,
-        List<PlayerDto> players
-) {
-}
