--- src/main/java/academy/cheerlot/repository/GameRepository.java
@@ -1,7 +1,0 @@
-package academy.cheerlot.repository;
-
-import academy.cheerlot.domain.Game;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface GameRepository extends JpaRepository<Game, String> {
-}
