--- src/main/java/academy/cheerlot/repository/TeamRepository.java
@@ -1,7 +1,0 @@
-package academy.cheerlot.repository;
-
-import academy.cheerlot.domain.Team;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface TeamRepository extends JpaRepository<Team, String> {
-}
