+++ src/main/java/academy/cheerlot/shared/config/RestTemplateConfig.java
@@ -1,0 +1,14 @@
+package academy.cheerlot.shared.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.web.client.RestTemplate;
+
+@Configuration
+public class RestTemplateConfig {
+
+    @Bean
+    public RestTemplate restTemplate() {
+        return new RestTemplate();
+    }
+}
